Creacion de clase persona
