+++ Git/src/Main.java
@@ -1,19 +1,6 @@
+import Persona.Persona;
+
 public class Main {
-
-    static class Persona{
-        String nombre;
-        String pasatiempo;
-
-        public  Persona( String nombre, String pasatiempo){
-            this.nombre = nombre;
-            this.pasatiempo = pasatiempo;
-        }
-
-        public void presentacion(){
-            System.out.println(" Hola soy " + nombre + " y me gusta mucho " + pasatiempo );
-        }
-
-    }
 
     public static void main(String[] args) {
